Replace HttpClientIntegration with HttpWebInvokerIntegration which allows distributed-tracing headers to be written
